requirements.txt
fastapi
uvicorn
aiofiles
python-multipart
requests
python-docx
pytesseract
pdfplumber
Pillow
jinja2